Visitor moved to command package
