+++ src/main/java/edu/kis/powp/jobs2d/command/SetPositionCommand.java
@@ -1,7 +1,7 @@
 package edu.kis.powp.jobs2d.command;
 
 import edu.kis.powp.jobs2d.Job2dDriver;
-import edu.kis.powp.jobs2d.visitor.CommandVisitor;
+import edu.kis.powp.jobs2d.command.visitor.CommandVisitor;
 
 /**
  * Implementation of Job2dDriverCommand for setPosition command functionality.
